Refactor get_user and list_users methods.
