+++ lib/flix_web/controllers/fan_controller.ex
@@ -1,15 +1,15 @@
 defmodule FlixWeb.FanController do
   use FlixWeb, :controller
 
-  alias Flix.Catalogs
+  alias Flix.Accounts
 
   def index(conn, _params) do
-    users = Catalogs.list_users()
+    users = Accounts.list_users()
     render(conn, "index.html", users: users)
   end
 
   def show(conn, %{"id" => id}) do
-    user = Catalogs.get_user!(id) 
+    user = Accounts.get_user!(id)
     reviews = user.reviews
     favorite_movies = user.favorite_movies
